# [RNMobile] Performance improvements in LatestPosts related controls