--- a/packages/components/src/query-controls/category-select.js
+++ b/packages/components/src/query-controls/category-select.js
@@ -3,6 +3,10 @@
  */
 import { buildTermsTree } from './terms';
 import TreeSelect from '../tree-select';
+/**
+ * WordPress dependencies
+ */
+import { useMemo } from '@wordpress/element';
 
 export default function CategorySelect( {
 	label,
@@ -12,7 +16,10 @@ export default function CategorySelect( {
 	onChange,
 	...props
 } ) {
-	const termsTree = buildTermsTree( categoriesList );
+	const termsTree = useMemo( () => {
+		return buildTermsTree( categoriesList );
+	}, [ categoriesList ] );
+
 	return (
 		<TreeSelect
 			{ ...{ label, noOptionLabel, onChange } }