--- a/packages/components/src/range-control/index.native.js
+++ b/packages/components/src/range-control/index.native.js
@@ -1,64 +1,70 @@
 /**
+ * External dependencies
+ */
+import React from 'react';
+/**
  * Internal dependencies
  */
 import RangeCell from '../mobile/bottom-sheet/range-cell';
 import StepperCell from '../mobile/bottom-sheet/stepper-cell';
 
-function RangeControl( {
-	className,
-	currentInput,
-	label,
-	value,
-	instanceId,
-	onChange,
-	beforeIcon,
-	afterIcon,
-	help,
-	allowReset,
-	initialPosition,
-	min,
-	max,
-	type,
-	separatorType,
-	...props
-} ) {
-	if ( type === 'stepper' ) {
+const RangeControl = React.memo(
+	( {
+		className,
+		currentInput,
+		label,
+		value,
+		instanceId,
+		onChange,
+		beforeIcon,
+		afterIcon,
+		help,
+		allowReset,
+		initialPosition,
+		min,
+		max,
+		type,
+		separatorType,
+		...props
+	} ) => {
+		if ( type === 'stepper' ) {
+			return (
+				<StepperCell
+					label={ label }
+					max={ max }
+					min={ min }
+					onChange={ onChange }
+					separatorType={ separatorType }
+					value={ value }
+				/>
+			);
+		}
+		const id = `inspector-range-control-${ instanceId }`;
+		const currentInputValue = currentInput === null ? value : currentInput;
+		const initialSliderValue = isFinite( currentInputValue )
+			? currentInputValue
+			: initialPosition;
+
 		return (
-			<StepperCell
+			<RangeCell
 				label={ label }
-				max={ max }
-				min={ min }
+				id={ id }
+				help={ help }
+				className={ className }
 				onChange={ onChange }
-				separatorType={ separatorType }
+				aria-describedby={ !! help ? `${ id }__help` : undefined }
+				minimumValue={ min }
+				maximumValue={ max }
 				value={ value }
+				beforeIcon={ beforeIcon }
+				afterIcon={ afterIcon }
+				allowReset={ allowReset }
+				defaultValue={ initialSliderValue }
+				separatorType={ separatorType }
+				{ ...props }
 			/>
 		);
 	}
-	const id = `inspector-range-control-${ instanceId }`;
-	const currentInputValue = currentInput === null ? value : currentInput;
-	const initialSliderValue = isFinite( currentInputValue )
-		? currentInputValue
-		: initialPosition;
-
-	return (
-		<RangeCell
-			label={ label }
-			id={ id }
-			help={ help }
-			className={ className }
-			onChange={ onChange }
-			aria-describedby={ !! help ? `${ id }__help` : undefined }
-			minimumValue={ min }
-			maximumValue={ max }
-			value={ value }
-			beforeIcon={ beforeIcon }
-			afterIcon={ afterIcon }
-			allowReset={ allowReset }
-			defaultValue={ initialSliderValue }
-			separatorType={ separatorType }
-			{ ...props }
-		/>
-	);
-}
+);
 
 export default RangeControl;