--- a/packages/components/src/toggle-control/index.native.js
+++ b/packages/components/src/toggle-control/index.native.js
@@ -1,31 +1,29 @@
 /**
+ * External dependencies
+ */
+import React from 'react';
+/**
  * Internal dependencies
  */
 import SwitchCell from '../mobile/bottom-sheet/switch-cell';
 
-function ToggleControl( {
-	label,
-	checked,
-	help,
-	instanceId,
-	className,
-	onChange,
-	...props
-} ) {
-	const id = `inspector-toggle-control-${ instanceId }`;
+const ToggleControl = React.memo(
+	( { label, checked, help, instanceId, className, onChange, ...props } ) => {
+		const id = `inspector-toggle-control-${ instanceId }`;
 
-	return (
-		<SwitchCell
-			label={ label }
-			id={ id }
-			help={ help }
-			className={ className }
-			value={ checked }
-			onValueChange={ onChange }
-			aria-describedby={ !! help ? id + '__help' : undefined }
-			{ ...props }
-		/>
-	);
-}
+		return (
+			<SwitchCell
+				label={ label }
+				id={ id }
+				help={ help }
+				className={ className }
+				value={ checked }
+				onValueChange={ onChange }
+				aria-describedby={ !! help ? id + '__help' : undefined }
+				{ ...props }
+			/>
+		);
+	}
+);
 
 export default ToggleControl;