<!-- Learn the overall process and best practices for pull requests at https://github.com/WordPress/gutenberg/blob/master/docs/contributors/repository-management.md#pull-requests. -->

Part of: https://github.com/WordPress/gutenberg/pull/28196

Contains changes only related to the `Latest Posts` block and controls available within block settings.

| BLOCK			| KIND OF SETTING		| CONTROL COMPONENT	| NESTED CONTROL COMPONENT	|
| :---			| :---: 				|  :---: 			| :---:  					|
| LatestPosts   | Post Content Settings	| `ToggleControl`   | 	 						|
|   			|      					| `ToggleControl`   | 							|	
|   			|      					| `RangeControl`    | 							|
|  				| Post Meta Settings    | `ToggleControl`   |  							|
|   			| Sorting and filtering | `QueryControls`   | `SelectControl`			|
|   			|  						| 					| `CategoryControl`			|
|   			|  						| 					| `RangeControl`			|


## How has this been tested?
<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, tests ran to see how -->
<!-- your change affects other areas of the code, etc. -->

Test manually `LatestPost` block on mobile and web.

1. Open the app
2. Add `LatestPosts` block
3. Open block settings
4. Play with settings and confirm that all of them works
5. Keep all the options enabled
6. Save the post and open it on web

**Expect all the enabled options are displayed properly**

## Screenshots <!-- if applicable -->

### RESULTS

**1. toggling show post content**

- updating and mounting time reduced from **16.54ms** to **6.68ms**
- avoid redundant updates in other controls eg `QueryControl`

before | after
--- | ---
<img width="1137" alt="toggle_show_post_before" src="https://user-images.githubusercontent.com/22746080/104640854-532f4280-56a9-11eb-88e9-01b3a5206cca.png"> | <img width="1137" alt="toggle_show_post" src="https://user-images.githubusercontent.com/22746080/104640846-4f9bbb80-56a9-11eb-8b3b-dbdba60e5f15.png">

**2. Toggling post date when post content is presented**

- updating time reduce from **18.79ms** to **2.61ms**
- avoid redundant updates in other controls

before | after
--- | ---
<img width="1137" alt="toggle_display_post_date_before" src="https://user-images.githubusercontent.com/22746080/104641816-8e7e4100-56aa-11eb-9cc0-a1e04690437b.png"> | <img width="1137" alt="toggle_display_post_date" src="https://user-images.githubusercontent.com/22746080/104641824-90e09b00-56aa-11eb-9d78-e0acf0c8ee83.png">

**3. Change number of items**

- updating time reduce from **12.83ms** to **3.66ms**
- avoid redundant updates in other controls

before | after
--- | ---
<img width="1137" alt="trigger_num_of_items_before" src="https://user-images.githubusercontent.com/22746080/104642744-9c809180-56ab-11eb-9893-7784c4586b15.png"> | <img width="1137" alt="trigger_num_of_items" src="https://user-images.githubusercontent.com/22746080/104642757-a0141880-56ab-11eb-9d08-86aabb5e5534.png">




## Types of changes
<!-- What types of changes does your code introduce?  -->
<!-- Bug fix (non-breaking change which fixes an issue) -->
<!-- New feature (non-breaking change which adds functionality) -->
<!-- Breaking change (fix or feature that would cause existing functionality to not work as expected) -->

Performance improvements

## Checklist:
- [x] My code is tested.
- [x] My code follows the WordPress code style. <!-- Check code: `npm run lint`, Guidelines: https://developer.wordpress.org/coding-standards/wordpress-coding-standards/javascript/ -->
- [x] My code follows the accessibility standards. <!-- Guidelines: https://developer.wordpress.org/coding-standards/wordpress-coding-standards/accessibility/ -->
- [x] My code has proper inline documentation. <!-- Guidelines: https://developer.wordpress.org/coding-standards/inline-documentation-standards/javascript/ -->
- [x] I've included developer documentation if appropriate. <!-- Handbook: https://developer.wordpress.org/block-editor/ -->
- [x] I've updated all React Native files affected by any refactorings/renamings in this PR. <!-- React Native mobile Gutenberg guidelines: https://github.com/WordPress/gutenberg/blob/master/docs/contributors/native-mobile.md -->